Update const.py to expose add'l movement variables
I've added movement and movement_bucket to the list of variables exposed by default by the API.  This will allow for mimicking of the Owlet app by using the movement value to indicate the amount of movement.  I left the boolean value of moving so it does not break previous implementations, but I think if you wanted to convert the movement integer to a boolean, it should be done on the application side and not the api.

Also, wondering if the api should not expose ALL of the variables it has available to it and choose what you ignore in the application code as well.  This would reduce requests for implementation of other variables by other apps and leave your pyowletapi module as the default module for most applications that need to access their owlet data.